Extract yAxisLabel to separate file
