+++ src/components/charts/chart-composed/chart-composed-styles.scss
@@ -1,9 +1,9 @@
 @import '~styles/settings.scss';
 
 .legend {
-    min-height: 30px;
+  min-height: 30px;
 }
 
 .wrapper * {
-    font-family: $font-family-1;
+  font-family: $font-family-1;
 }
