+++ src/components/charts/chart/chart-styles.scss
@@ -12,6 +12,6 @@
   min-height: 30px;
 }
 
-.wrapper *{
+.wrapper * {
   font-family: $font-family-1;
 }
