+++ src/components/charts/line/line.js
@@ -9,13 +9,13 @@ import {
   CartesianGrid,
   Tooltip,
   ResponsiveContainer,
-  Label,
   ReferenceArea
 } from 'recharts';
-import { getMaxValue, htmlToSvgSubscript } from 'utils';
+import { getMaxValue } from 'utils';
 import TooltipChart from 'components/charts/tooltip-chart';
 import debounce from 'lodash/debounce';
 import isUndefined from 'lodash/isUndefined';
+import yAxisLabel from 'components/charts/y-axis-label';
 import DividerLine from '../projected-data/divider-line';
 import ProjectedData from '../projected-data';
 import { CustomXAxisTick, CustomYAxisTick } from './axis-ticks';
@@ -88,17 +88,7 @@ class ChartLine extends PureComponent {
     const dataMaxMin = getDataMaxMin(lineState);
     const domain = projectedData ? getDomain(lineState) : customDomain;
     const lastData = getMaxValue(getDataWithTotal(lineState));
-    const yAxisLabel = (
-      <Label
-        position="top"
-        offset={20}
-        content={() => (
-          <text x="8" y="20" fontSize="13px" stroke="#b1b1c1" strokeWidth="0.5">
-            {htmlToSvgSubscript(unit)}
-          </text>
-        )}
-      />
-    );
+
     return (
       <ResponsiveContainer
         height={height}
@@ -139,7 +129,7 @@ class ChartLine extends PureComponent {
             domain={domain && domain.y || [ 'auto', 'auto' ]}
             interval="preserveStartEnd"
           >
-            {yAxisLabel}
+            {yAxisLabel(unit)}
           </YAxis>
           <CartesianGrid vertical={false} />
           {
