+++ src/components/charts/stacked-area/stacked-area-styles.scss
